requirements.txt
typing~=3.7.4.3
setuptools~=67.8.0
click~=8.1.3
summ-eval~=0.892
pyyaml~=6.0
scipy~=1.10.1
huspacy~=0.9.0
six~=1.16.0
torch~=1.13.1
transformers~=4.30.2
pandas~=2.0.2
blanc~=0.3.0
gensim~=4.3.2

nltk~=3.8.1
numpy~=1.25.0
pyemd~=0.5.1
fasttext~=0.9.2
seaborn~=0.12.2